Remove all stuff between dotnet-compile and dotnet-compile-csc.
