+++ src/Orchard.Environment.Extensions/Compilers/CSharpExtensionCompiler.cs
@@ -1,0 +1,276 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Runtime.InteropServices;
+using Microsoft.DotNet.Cli.Compiler.Common;
+using Microsoft.DotNet.Cli.Utils;
+using Microsoft.DotNet.ProjectModel;
+using Microsoft.DotNet.ProjectModel.Compilation;
+using Microsoft.DotNet.ProjectModel.Files;
+using Microsoft.Extensions.DependencyModel;
+using NuGet.Frameworks;
+
+namespace Orchard.Environment.Extensions.Compilers
+{
+    public interface ICompiler
+    {
+        bool Compile(ProjectContext context, string config, string buildBasePath);
+    }
+
+    public class CSharpExtensionCompiler : ICompiler
+    {
+        public bool Compile(ProjectContext context, string config, string buildBasePath)
+        {
+            // Set up Output Paths
+            var outputPaths = context.GetOutputPaths(config, buildBasePath);
+            var outputPath = outputPaths.CompilationOutputPath;
+            var intermediateOutputPath = outputPaths.IntermediateOutputDirectoryPath;
+
+            Directory.CreateDirectory(outputPath);
+            Directory.CreateDirectory(intermediateOutputPath);
+
+            // Create the library exporter
+            var exporter = context.CreateExporter(config, buildBasePath);
+
+            // Gather exports for the project
+            var dependencies = exporter.GetDependencies().ToList();
+
+            var diagnostics = new List<DiagnosticMessage>();
+            var missingFrameworkDiagnostics = new List<DiagnosticMessage>();
+
+            // Collect dependency diagnostics
+            foreach (var diag in context.LibraryManager.GetAllDiagnostics())
+            {
+                if (diag.ErrorCode == ErrorCodes.DOTNET1011 ||
+                    diag.ErrorCode == ErrorCodes.DOTNET1012)
+                {
+                    missingFrameworkDiagnostics.Add(diag);
+                }
+
+                diagnostics.Add(diag);
+            }
+
+            if(diagnostics.Any(d => d.Severity == DiagnosticMessageSeverity.Error))
+            {
+                // We got an unresolved dependency or missing framework. Don't continue the compilation.
+                return false;
+            }
+
+            // Get compilation options
+            var outputName = outputPaths.CompilationFiles.Assembly;
+            var compilationOptions = context.ResolveCompilationOptions(config);
+
+            // Set default platform if it isn't already set and we're on desktop
+            if (compilationOptions.EmitEntryPoint == true && string.IsNullOrEmpty(compilationOptions.Platform) && context.TargetFramework.IsDesktop())
+            {
+                // See https://github.com/dotnet/cli/issues/2428 for more details.
+                compilationOptions.Platform = RuntimeInformation.ProcessArchitecture == Architecture.X64 ?
+                    "x64" : "anycpu32bitpreferred";
+            }
+
+            var references = new List<string>();
+            var sourceFiles = new List<string>();
+
+            // Add metadata options
+            var assemblyInfoOptions = AssemblyInfoOptions.CreateForProject(context);
+
+            foreach (var dependency in dependencies)
+            {
+                references.AddRange(dependency.CompilationAssemblies.Select(r => r.ResolvedPath));
+                sourceFiles.AddRange(dependency.SourceReferences.Select(s => s.GetTransformedFile(intermediateOutputPath)));
+            }
+
+            var resources = new List<string>();
+            if (compilationOptions.PreserveCompilationContext == true)
+            {
+                var allExports = exporter.GetAllExports().ToList();
+                var dependencyContext = new DependencyContextBuilder().Build(compilationOptions,
+                    allExports,
+                    allExports,
+                    false, // For now, just assume non-portable mode in the legacy deps file (this is going away soon anyway)
+                    context.TargetFramework,
+                    context.RuntimeIdentifier ?? string.Empty);
+
+                var writer = new DependencyContextWriter();
+                var depsJsonFile = Path.Combine(intermediateOutputPath, compilationOptions.OutputName + "dotnet-compile.deps.json");
+                using (var fileStream = File.Create(depsJsonFile))
+                {
+                    writer.Write(dependencyContext, fileStream);
+                }
+
+                resources.Add($"\"{depsJsonFile}\",{compilationOptions.OutputName}.deps.json");
+            }
+
+            // Add project source files
+            if (compilationOptions.CompileInclude == null)
+            {
+                sourceFiles.AddRange(context.ProjectFile.Files.SourceFiles);
+            }
+            else {
+                var includeFiles = IncludeFilesResolver.GetIncludeFiles(compilationOptions.CompileInclude, "/", diagnostics: null);
+                sourceFiles.AddRange(includeFiles.Select(f => f.SourcePath));
+            }
+
+            if (String.IsNullOrEmpty(intermediateOutputPath))
+            {
+                return false;
+            }
+
+            var translated = TranslateCommonOptions(compilationOptions, outputName);
+
+            var allArgs = new List<string>(translated);
+            allArgs.AddRange(GetDefaultOptions());
+
+            // Generate assembly info
+            var assemblyInfo = Path.Combine(intermediateOutputPath, $"dotnet-compile.assemblyinfo.cs");
+
+            File.WriteAllText(assemblyInfo, AssemblyInfoFileGenerator.GenerateCSharp(assemblyInfoOptions, sourceFiles));
+
+            allArgs.Add($"\"{assemblyInfo}\"");
+
+            if (!String.IsNullOrEmpty(outputName))
+            {
+                allArgs.Add($"-out:\"{outputName}\"");
+            }
+
+            allArgs.AddRange(references.Select(r => $"-r:\"{r}\""));
+            allArgs.AddRange(resources.Select(resource => $"-resource:{resource}"));
+            allArgs.AddRange(sourceFiles.Select(s => $"\"{s}\""));
+            allArgs.Prepend($"-noconfig");
+
+            // Execute CSC!
+            var result = RunCsc(allArgs.ToArray())
+                .WorkingDirectory(Directory.GetCurrentDirectory())
+                .ForwardStdErr()
+                .ForwardStdOut()
+                .Execute();
+
+            return result.ExitCode == 0;
+        }
+
+        // TODO: Review if this is the place for default options
+        private static IEnumerable<string> GetDefaultOptions()
+        {
+            var args = new List<string>()
+            {
+                "-nostdlib",
+                "-nologo",
+            };
+
+            return args;
+        }
+
+        private static IEnumerable<string> TranslateCommonOptions(CommonCompilerOptions options, string outputName)
+        {
+            List<string> commonArgs = new List<string>();
+
+            if (options.Defines != null)
+            {
+                commonArgs.AddRange(options.Defines.Select(def => $"-d:{def}"));
+            }
+
+            if (options.SuppressWarnings != null)
+            {
+                commonArgs.AddRange(options.SuppressWarnings.Select(w => $"-nowarn:{w}"));
+            }
+
+            // Additional arguments are added verbatim
+            if (options.AdditionalArguments != null)
+            {
+                commonArgs.AddRange(options.AdditionalArguments);
+            }
+
+            if (options.LanguageVersion != null)
+            {
+                commonArgs.Add($"-langversion:{GetLanguageVersion(options.LanguageVersion)}");
+            }
+
+            if (options.Platform != null)
+            {
+                commonArgs.Add($"-platform:{options.Platform}");
+            }
+
+            if (options.AllowUnsafe == true)
+            {
+                commonArgs.Add("-unsafe");
+            }
+
+            if (options.WarningsAsErrors == true)
+            {
+                commonArgs.Add("-warnaserror");
+            }
+
+            if (options.Optimize == true)
+            {
+                commonArgs.Add("-optimize");
+            }
+
+            if (options.KeyFile != null)
+            {
+                commonArgs.Add($"-keyfile:\"{options.KeyFile}\"");
+
+                // If we're not on Windows, full signing isn't supported, so we'll
+                // public sign, unless the public sign switch has explicitly been
+                // set to false
+                if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows) &&
+                    options.PublicSign == null)
+                {
+                    commonArgs.Add("-publicsign");
+                }
+            }
+
+            if (options.DelaySign == true)
+            {
+                commonArgs.Add("-delaysign");
+            }
+
+            if (options.PublicSign == true)
+            {
+                commonArgs.Add("-publicsign");
+            }
+
+            if (options.GenerateXmlDocumentation == true)
+            {
+                commonArgs.Add($"-doc:\"{Path.ChangeExtension(outputName, "xml")}\"");
+            }
+
+            if (options.EmitEntryPoint != true)
+            {
+                commonArgs.Add("-t:library");
+            }
+
+            if (string.IsNullOrEmpty(options.DebugType))
+            {
+                commonArgs.Add(RuntimeInformation.IsOSPlatform(OSPlatform.Windows)
+                    ? "-debug:full"
+                    : "-debug:portable");
+            }
+            else
+            {
+                commonArgs.Add(options.DebugType == "portable"
+                    ? "-debug:portable"
+                    : "-debug:full");
+            }
+
+            return commonArgs;
+        }
+
+        private static string GetLanguageVersion(string languageVersion)
+        {
+            // project.json supports the enum that the roslyn APIs expose
+            if (languageVersion?.StartsWith("csharp", StringComparison.OrdinalIgnoreCase) == true)
+            {
+                // We'll be left with the number csharp6 = 6
+                return languageVersion.Substring("csharp".Length);
+            }
+            return languageVersion;
+        }
+
+        private static Command RunCsc(string[] cscArgs)
+        {
+            // Locate CoreRun
+            return Command.Create("csc.dll", cscArgs);
+        }
+    }
+}
