--- src/Orchard.Environment.Extensions/Compilers/CommandArgument.cs
@@ -1,29 +1,0 @@
-// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace Microsoft.DotNet.Cli.CommandLine
-{
-    internal class CommandArgument
-    {
-        public CommandArgument()
-        {
-            Values = new List<string>();
-        }
-
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public List<string> Values { get; private set; }
-        public bool MultipleValues { get; set; }
-        public string Value
-        {
-            get
-            {
-                return Values.FirstOrDefault();
-            }
-        }
-    }
-}
