--- src/Orchard.Environment.Extensions/Compilers/CommandOptionType.cs
@@ -1,14 +1,0 @@
-// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-
-namespace Microsoft.DotNet.Cli.CommandLine
-{
-    internal enum CommandOptionType
-    {
-        MultipleValue,
-        SingleValue,
-        BoolValue,
-        NoValue
-    }
-}
