--- src/Orchard.Environment.Extensions/Compilers/CommandParsingException.cs
@@ -1,18 +1,0 @@
-// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System;
-
-namespace Microsoft.DotNet.Cli.CommandLine
-{
-    internal class CommandParsingException : Exception
-    {
-        public CommandParsingException(CommandLineApplication command, string message)
-            : base(message)
-        {
-            Command = command;
-        }
-
-        public CommandLineApplication Command { get; }
-    }
-}
