--- src/Orchard.Environment.Extensions/Compilers/Compiler.cs
@@ -1,128 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Linq;
-using Microsoft.DotNet.Cli.Utils;
-using Microsoft.DotNet.ProjectModel;
-using Microsoft.DotNet.ProjectModel.Compilation;
-
-namespace Microsoft.DotNet.Tools.Compiler
-{
-    public interface ICompiler
-    {
-        bool Compile(ProjectContext context, string config, string buildBasePath);
-    }
-
-    public abstract class Compiler : ICompiler
-    {
-        public abstract bool Compile(ProjectContext context, string config, string buildBasePath);
-
-        protected static bool PrintSummary(List<DiagnosticMessage> diagnostics, Stopwatch sw, bool success = true)
-        {
-            PrintDiagnostics(diagnostics);
-
-            Reporter.Output.WriteLine();
-
-            var errorCount = diagnostics.Count(d => d.Severity == DiagnosticMessageSeverity.Error);
-            var warningCount = diagnostics.Count(d => d.Severity == DiagnosticMessageSeverity.Warning);
-
-            if (errorCount > 0 || !success)
-            {
-                Reporter.Output.WriteLine("Compilation failed.".Red());
-                success = false;
-            }
-            else
-            {
-                Reporter.Output.WriteLine("Compilation succeeded.".Green());
-            }
-
-            Reporter.Output.WriteLine($"    {warningCount} Warning(s)");
-            Reporter.Output.WriteLine($"    {errorCount} Error(s)");
-
-            Reporter.Output.WriteLine();
-
-            Reporter.Output.WriteLine($"Time elapsed {sw.Elapsed}");
-
-            return success;
-        }
-
-        protected static DiagnosticMessage ParseDiagnostic(string projectRootPath, string line)
-        {
-            var error = CanonicalError.Parse(line);
-
-            if (error != null)
-            {
-                var severity = error.category == CanonicalError.Parts.Category.Error ?
-                DiagnosticMessageSeverity.Error : DiagnosticMessageSeverity.Warning;
-
-                return new DiagnosticMessage(
-                    error.code,
-                    error.text,
-                    Path.IsPathRooted(error.origin) ? line : projectRootPath + Path.DirectorySeparatorChar + line,
-                    Path.Combine(projectRootPath, error.origin),
-                    severity,
-                    error.line,
-                    error.column,
-                    error.endColumn,
-                    error.endLine,
-                    source: null);
-            }
-
-            return null;
-        }
-
-        private static void PrintDiagnostics(List<DiagnosticMessage> diagnostics)
-        {
-            foreach (var diag in diagnostics)
-            {
-                PrintDiagnostic(diag);
-            }
-        }
-
-        private static void PrintDiagnostic(DiagnosticMessage diag)
-        {
-            switch (diag.Severity)
-            {
-                case DiagnosticMessageSeverity.Info:
-                    Reporter.Error.WriteLine(diag.FormattedMessage);
-                    break;
-                case DiagnosticMessageSeverity.Warning:
-                    Reporter.Error.WriteLine(diag.FormattedMessage.Yellow().Bold());
-                    break;
-                case DiagnosticMessageSeverity.Error:
-                    Reporter.Error.WriteLine(diag.FormattedMessage.Red().Bold());
-                    break;
-            }
-        }
-
-        private static void CopyFiles(IEnumerable<LibraryAsset> files, string outputPath)
-        {
-            foreach (var file in files)
-            {
-                File.Copy(file.ResolvedPath, Path.Combine(outputPath, Path.GetFileName(file.ResolvedPath)), overwrite: true);
-            }
-        }
-
-        private static string EnsureTrailingSlash(string path)
-        {
-            return EnsureTrailingCharacter(path, Path.DirectorySeparatorChar);
-        }
-
-        private static string EnsureTrailingCharacter(string path, char trailingCharacter)
-        {
-            if (path == null)
-            {
-                throw new ArgumentNullException(nameof(path));
-            }
-
-            // if the path is empty, we want to return the original string instead of a single trailing character.
-            if (path.Length == 0 || path[path.Length - 1] == trailingCharacter)
-            {
-                return path;
-            }
-
-            return path + trailingCharacter;
-        }
-    }
-}
